[RTOP-69] Files remove and Compilation management
