+++ scripts/compile-clean.sh
@@ -1,0 +1,7 @@
+#!/bin/bash
+set -euo pipefail
+
+# Move result
+mkdir -p build
+mv libplc.so build/
+rm *.o
